Cleaned up metanode naming to reduce computation.
If a inner node has 0 flow, it is not included in the state, thus condensing
the variety of metanodes and resulting in fewer computations. MG_Custom1 went
from ~1.7MB to ~0.6MB. Still, too large...needs further optimizations, perhaps
by only allowing stationary flow at the target (and maybe start) to reduce the
large the number of intermediary metanodes.